+++ bc/recruitment/templatetags/jobs_search_filters.py
@@ -66,13 +66,13 @@ def jobs_search_filters(request, unfiltered_results=None):
             },
             {
                 "title": "Salary range",
-                "options": unfiltered_results.exclude(salary_range__exact="")
-                .values("salary_range")
-                .annotate(key=F("salary_range"), label=F("salary_range"))
-                .order_by("salary_range")
+                "options": unfiltered_results.exclude(searchable_salary__exact="")
+                .values("searchable_salary")
+                .annotate(key=F("searchable_salary"), label=F("searchable_salary"))
+                .order_by("searchable_salary")
                 .distinct(),
-                "selected": request.GET.getlist("salary_range"),
-                "key": "salary_range",
+                "selected": request.GET.getlist("searchable_salary"),
+                "key": "searchable_salary",
             },
         ],
         "search_postcode": search_postcode,